# lib/api-zod/src/generated/types/antiNukeConfig.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AntiNukeConfig {
  enabled: boolean;
  /** @nullable */
  baitChannelId: string | null;
  /** @minimum 1 */
  firstTimeoutHours: number;
  /** @minimum 1 */
  secondTimeoutDays: number;
  /** @nullable */
  mutedRoleId: string | null;
  purgeRecentMessages: boolean;
  dmReason: boolean;
}