# lib/api-zod/src/generated/types/discordGuild.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiscordGuild {
  id: string;
  name: string;
  memberCount: number;
  /** @nullable */
  icon: string | null;
}