# lib/api-zod/src/generated/types/embedField.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmbedField {
  name: string;
  value: string;
  inline: boolean;
}