# lib/api-zod/src/generated/types/embedTemplate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmbedField } from './embedField';

export interface EmbedTemplate {
  guildId: string;
  key: string;
  name: string;
  description: string;
  color: string;
  content: string;
  fields: EmbedField[];
  footer: string;
  enabled: boolean;
}